+++ ScaryFlight/ScaryFlight/Classes/Sprites/EAHero.m
@@ -17,7 +17,7 @@ static CGFloat const kHeroDirection = 28.5f;
     NSParameterAssert(yLimit > 0.0f);
     
     if (self.position.y < yLimit - self.size.height / 2.0f) { // <-- avoid hero to fly away from top of screen
-        CGFloat heroDirection = self.zRotation + M_PI_2;
+        float heroDirection = self.zRotation + (float)M_PI_2;
         self.physicsBody.velocity = CGVectorMake(0.0f, 0.0f);
         [self.physicsBody applyImpulse:CGVectorMake(kHeroDirection * cosf(heroDirection),
                                                     kHeroDirection * sinf(heroDirection))];
